adding tutorial link to SemanticDictionary.ipynb
resolves tensorflow/lucid#43

Hotlinked the text "Check out" to the following URL:
- https://colab.sandbox.google.com/github/tensorflow/lucid/blob/master/notebooks/tutorial.ipynb

Note: as a non-googler i get redirected to the following URL, would this
be more appropriate?
- https://colab.research.google.com/github/tensorflow/lucid/blob/master/notebooks/tutorial.ipynb

One question: I noticed that when i run jupyter locally to edit the
notebook, it made the ipynb JSON single-space-indented instead of
double. I'm assuming colab's fork of jupyter deals with the ipynb file
slightly differently? If this is important to maintainers, I'm open to
alternatives (maybe building & running `googlecolab/backend-container`
docker image would be a safe bet? I'd be happy to add a docs section
on how to run this repo's notebooks inside a docker instance, if we
think that might make it easier for other developers to get started.)