tests: save to temporary directory
There were two tests that wrote to the same file in tests/fixtures
which lead to the annoyance of manually having to revert the file
to avoid spurious commit changes. Probably the order of test
execution isn't stable.

With this change, the pytest 'tmpdir' fixture is used to provide a
unique temporary directory for each test.